FloatField Changed to DecimalField in all models

+++ bet_bola/history/models.py
@@ -9,9 +9,9 @@ class SellerSalesHistory(models.Model):
     seller = models.ForeignKey(Seller, on_delete=models.CASCADE, verbose_name='Vendedor')
     bet_ticket = models.ForeignKey(BetTicket, on_delete=models.CASCADE, verbose_name='Ticket Criado')
     sell_date = models.DateTimeField(verbose_name='Data da Venda', auto_now_add=True)
-    value = models.FloatField(verbose_name='Valor Apostado')
-    seller_before_balance = models.FloatField(null=True,blank=True, verbose_name='Saldo Anterior')
-    seller_after_balance = models.FloatField(null=True, blank=True, verbose_name='Saldo Atual')
+    value = models.DecimalField(max_digits=30, decimal_places=2,verbose_name='Valor Apostado')
+    seller_before_balance = models.DecimalField(max_digits=30, decimal_places=2,null=True,blank=True, verbose_name='Saldo Anterior')
+    seller_after_balance = models.DecimalField(max_digits=30, decimal_places=2,null=True, blank=True, verbose_name='Saldo Atual')
 
     def __str__(self):
         return "Transações - Vendedores"
@@ -27,11 +27,11 @@ class ManagerTransactions(models.Model):
     manager = models.ForeignKey(Manager, on_delete=models.CASCADE, verbose_name='Gerente')
     seller = models.ForeignKey(Seller, on_delete=models.CASCADE, verbose_name='Vendedor')
     transaction_date = models.DateTimeField(verbose_name='Data da Transação', auto_now_add=True)
-    transferred_amount = models.FloatField(verbose_name='Valor Transferido')
-    manager_before_balance = models.FloatField(null=True,blank=True, verbose_name='Saldo Anterior')
-    manager_after_balance = models.FloatField(null=True,blank=True, verbose_name='Saldo Atual')
-    seller_before_balance = models.FloatField(null=True,blank=True, verbose_name='Saldo Anterior(Vendedor)')
-    seller_after_balance = models.FloatField(null=True,blank=True, verbose_name='Saldo Atual(Vendedor)')
+    transferred_amount = models.DecimalField(max_digits=30, decimal_places=2,verbose_name='Valor Transferido')
+    manager_before_balance = models.DecimalField(max_digits=30, decimal_places=2,null=True,blank=True, verbose_name='Saldo Anterior')
+    manager_after_balance = models.DecimalField(max_digits=30, decimal_places=2,null=True,blank=True, verbose_name='Saldo Atual')
+    seller_before_balance = models.DecimalField(max_digits=30, decimal_places=2,null=True,blank=True, verbose_name='Saldo Anterior(Vendedor)')
+    seller_after_balance = models.DecimalField(max_digits=30, decimal_places=2,null=True,blank=True, verbose_name='Saldo Atual(Vendedor)')
 
 
     def __str__(self):
@@ -46,9 +46,9 @@ class RevenueHistorySeller(models.Model):
     who_reseted_revenue = models.CharField(max_length=200, verbose_name='Reponsável pelo Fechamento')
     seller = models.ForeignKey(Seller, on_delete=models.CASCADE, verbose_name='Vendedor')
     revenue_reseted_date = models.DateTimeField(verbose_name='Data da Transação', auto_now_add=True)
-    final_revenue = models.FloatField(null=True, verbose_name='Faturamento')
-    actual_comission = models.FloatField(null=True, verbose_name='Comissão')
-    earned_value = models.FloatField(null=True, verbose_name='Valor Recebido')
+    final_revenue = models.DecimalField(max_digits=30, decimal_places=2,null=True, verbose_name='Faturamento')
+    actual_comission = models.DecimalField(max_digits=30, decimal_places=2,null=True, verbose_name='Comissão')
+    earned_value = models.DecimalField(max_digits=30, decimal_places=2,null=True, verbose_name='Valor Recebido')
 
     def __str__(self):
         return "Pagamentos - Vendedores"
@@ -63,9 +63,9 @@ class RevenueHistoryManager(models.Model):
     who_reseted_revenue = models.CharField(max_length=200, verbose_name='Reponsável pelo Fechamento')
     manager = models.ForeignKey(Manager, on_delete=models.CASCADE, verbose_name='Gerente')
     revenue_reseted_date = models.DateTimeField(verbose_name='Data da Transação', auto_now_add=True)
-    final_revenue = models.FloatField(null=True, verbose_name='Faturamento')
-    actual_comission = models.FloatField(null=True, verbose_name='Comissão')
-    earned_value = models.FloatField(null=True, verbose_name='Valor Recebido')
+    final_revenue = models.DecimalField(max_digits=30, decimal_places=2,null=True, verbose_name='Faturamento')
+    actual_comission = models.DecimalField(max_digits=30, decimal_places=2,null=True, verbose_name='Comissão')
+    earned_value = models.DecimalField(max_digits=30, decimal_places=2,null=True, verbose_name='Valor Recebido')
 
 
     def __str__(self):
@@ -81,7 +81,7 @@ class PunterPayedHistory(models.Model):
     seller = models.ForeignKey(Seller, on_delete=models.CASCADE, verbose_name='Vendedor')
     ticket_winner = models.ForeignKey(BetTicket, on_delete=models.CASCADE, verbose_name='Ticket Vencedor')
     payment_date = models.DateTimeField(verbose_name='Data do Pagamento', auto_now_add=True)
-    payed_value = models.FloatField(verbose_name='Valor Pago')
+    payed_value = models.DecimalField(max_digits=30, decimal_places=2,verbose_name='Valor Pago')
     is_closed = models.BooleanField(default=False)
 
     def __str__(self):
